Improved API of the analyzer for multilingual
