+++ lib/natural/sentiment/Spanish/negations.json
